refactor: stream_proxy dissolution stage 1 - extract helpers to 7 sibling modules (iter32)
First stage of decomposing the stream_proxy.py god-file: move ~90 module-level helper
functions + ReadAheadBuffer + the prepare/flush service client out into 7 cohesive
sibling modules (stream_proxy_{buffer,service,settings,ffmpeg,contract,recovery,fallback},
each 150-344 NLOC). stream_proxy.py NLOC 10,472 -> 5,959.

Behavior-preserving facade pattern: all moved names re-exported as stream_proxy.<name>;
siblings reach parent/patched names at call time via `import resources.lib.stream_proxy
as _sp`, preserving every @patch and adding no params. The three big classes
(_StreamHandler 4967, StreamProxy 1447, HlsProducer 567) stay in stream_proxy.py for now -
moving them whole would exceed 500 NLOC; they need mixin-splitting in later stages to
finally clear the file-nloc>500 flag.

Gauge held FLAT at 24 (no regression; this is a decomposition checkpoint, not yet a
reduction - the file-nloc flag clears once the classes are split). Verified: 2138 unit
tests pass, integration pass (real ffmpeg), pylint 10/10, ruff+black clean, vermin 3.8,
all new modules <500 NLOC.
